refactor queue handler
